Mensajes Sweet Alert 2 en Contactos y Ajustes CSS en registro de Ciudano y Actualizacion de Ciudadono

+++ client/admininfo/src/services/contactos/contactostable.js
@@ -2,6 +2,7 @@ import React from "react";
 import Button from "react-bootstrap/Button";
 
 import { useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
 
 import { BsFillPenFill, BsFillTrashFill } from "react-icons/bs";
 import { deleteCiudadano } from "../../routes/contactos";
@@ -16,18 +17,82 @@ const TableContactos = (props) => {
   const navigate = useNavigate();
 
   const borrarContacto = async () => {
-    let data = "";
-    const authheader = AuthHeaders();
 
-    try {
-      data = await deleteCiudadano(identification, authheader);
+        Swal.fire({
+          title: "Desea eliminar el usuario? ",
+          html: "Deleting will be canceled in 10 <strong></strong> seconds.",
+          timer: 10000,
+          timerProgressBar: true,
+          showDenyButton: true,
+          showConfirmButton: true,
+          confirmButtonText: "Delete",
+          denyButtonText: "Not delete",
+        }).then(async (result) => {
+          
 
-      console.log("Ciudadano Eliminado");
-      console.log(data);
-      window.location.reload();
-    } catch (error) {
-      console.log("Ciudadano no se ha podido eliminar", error);
-    }
+          try {
+
+            if (result.isConfirmed) {
+              const authheader = AuthHeaders();
+              await deleteCiudadano(identification, authheader);
+
+              Swal.fire({
+                icon: "success",
+                title: "Ciudadano Eliminado",
+                showConfirmButton: false,
+                timer: 2000,
+                didOpen: () => {
+                  Swal.showLoading();
+                },
+              });
+              setTimeout(() => {
+                Swal.close();
+                window.location.reload();
+              }, 2000);
+            } else {
+              if (result.isDenied) {
+                Swal.fire({
+                  icon: "info",
+                  title: "Ciudadano no ha sido Eliminado",
+                  showConfirmButton: false,
+                  timer: 2000,
+                  didOpen: () => {
+                    Swal.showLoading();
+                  },
+                });
+                setTimeout(() => {
+                  Swal.close();
+                }, 2000);
+              }
+            }
+            if (result.dismiss === Swal.DismissReason.timer) {
+              Swal.fire({
+                icon: "error",
+                title: "Se ha superado el tiempo sin una respuesta",
+                showConfirmButton: false,
+                timer: 2000,
+                didOpen: () => {
+                  Swal.showLoading();
+                },
+              });
+              setTimeout(() => {
+                Swal.close();
+              }, 2000);
+            }
+
+          } catch (error) {
+            let mensaje;
+            mensaje = error.response.data;
+
+            Swal.fire({
+              icon: "error",
+              title: mensaje,
+              showConfirmButton: false,
+              timer: 2000,
+            });
+            Swal.showLoading();
+          }
+        });
   };
 
   return (
@@ -49,7 +114,7 @@ const TableContactos = (props) => {
           </Button>
           <Button
             className="botones-mod"
-            variant="info"
+            variant="warning"
             onClick={() => navigate(`/contactos/${identification}`)}
           >
             <BsFillPenFill />
